% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAG_plot.R
\name{DAG_plot}
\alias{DAG_plot}
\title{Causal Mediation Analysis with Multiple Causally Ordered Mediators}
\arguments{
\item{va}{-- variable's name}

\item{pse_result}{-- results of PSE-MR}

\item{ylow}{-- lowest value of ylab}

\item{yup}{-- uppest value of ylab}

\item{tt}{-- label name}
}
\value{
list
}
\description{
Causal Mediation Analysis with Multiple Causally Ordered Mediators
}
