% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSE_MR_Egger_order.R
\name{Ordered_Egger}
\alias{Ordered_Egger}
\title{Causal Mediation Analysis with Multiple Causally Ordered Mediators}
\usage{
Ordered_Egger(betaYG, sebetaYG, betaXG, sebetaXG, betaMG, sebetaMG)
}
\arguments{
\item{betaYG}{-- a vector, genetic summary statistics for outcome (coefficient)}

\item{sebetaYG}{-- a vector, genetic summary statistics for outcome (standard error)}

\item{betaXG}{-- a vector, genetic summary statistics for exposure (coefficient)}

\item{sebetaXG}{-- a vector,genetic summary statistics for exposure (standard error)}

\item{betaMG}{-- matrix or dataframe with n mediators, genetic summary statistics for mediators (coefficient)}

\item{sebetaMG}{-- matrix or dataframe with n mediators, genetic summary statistics formediators (standard error)}
}
\value{
list
}
\description{
Causal Mediation Analysis with Multiple Causally Ordered Mediators
}
