% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Path_effect.R
\name{Path_effect}
\alias{Path_effect}
\title{Causal Mediation Analysis with Multiple Causally Ordered Mediators}
\usage{
Path_effect(coef_m, effect_m, se_m, straps)
}
\arguments{
\item{coef_m}{-- pathway martrix}

\item{effect_m}{-- estimated effect in Multi_IVW or Multi_MR_Egger}

\item{se_m}{-- estimated se in Multi_IVW or Multi_MR_Egger}

\item{straps}{-- repeated times for MCMC}
}
\description{
Causal Mediation Analysis with Multiple Causally Ordered Mediators
}
